backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
supabase==2.3.4
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
yt-dlp==2024.3.10
opencv-python-headless==4.9.0.80
numpy==1.26.4
requests==2.31.0
ffmpeg-python==0.2.0
python-dotenv==1.0.1
faiss-cpu==1.7.4
openai==1.12.0
huggingface-hub==0.20.3 